Use Q96 fixed point pricing (#27)
* Add CurrencyLibrary

* payable

* save changes

* compiles

* SSTORE2

* Add checkpointing

* adjust tests

* Add transferFrom

* Normalize to current bps

* Fix tests and gas

* forge fmt

* Add only update checkpoint test

* Save

* save

* Fix test math

* fix exact out test

* fix step advance

* Start step and previous end and fix parsing

* Add strategy and distribution contract interfaces

* Add multicall and permit2 forwarder

* add permit2

* Add permit2 to auction

* Move auction step to library

* Refactor for changes

* fmt

* natspec and nits

* add funcs to interface

* Gas savings

* Add exactOut fuzz test

* Add exactOut test

* msg.sender for safeTransferFrom

* Move bids into separate storage

* Refactor and add hook data

* forge fmt

* Add receive

* Remove payable from multicall

* Revert if multicall is used with msg.value

* rm unused files

* Fix imports

* emit owner with event

* fix conflicts with mps

* fix test

* rm multicall

* Add checkpoint storage

* Add getters and setters to bid storage

* rm multicall

* feedback

* Allow for withdraw before endBlock

* record step on construction

* Remove bid if no tokens filled

* Add checkpoint lib to simplify

* Add withdraw for bids at the clearing price

* fixes

* save

* New demand struct

* save

* pass all

* clean up

* renaming

* use fixed point math lib

* remove console2

* helper funcs

* Add comment

* simplfiy

* fmt

* cleanup

* clean up

* move transfer to end

* Only transfer if not 0

* remove v4-core

* isolate

* move transfer out of branch

* remove resolvedSupply

* fmt

* Use safecast

* simplify

* Refactor to upper tick

* Disallow bids below clearing price

* Checkpoint block cleared

* Implement correct partial fill calculation

* feat: restrict bids at clearing price (#22)

* Disallow bids at clearing price

* Fix tests

* Simplify partial fill calculation

* bid pricem ust be strictly greater than clearing price

* rmeove unused branching

* simplify

* fix

* fix blockCleared

* Add more test coverage (#15)

* Add withdrawBid test

* Add withdrawBid tests

* fix conflicts

* save

* fix

* Fix withdraw bid test now that can't bid at clearing price

* rm console2

* Clean up withdraw bid flow

* Add more tests

* Check for zero amounts

* Fix checkpointing by saving `prev` and clean up partial fill logic

* Account for tokens sold in the upper checkpoint block

* Add partial fill test

* save state

* works

* Add isolate

* comments

* Add another test

* fmt

* some optimizations

* fixup

* Add finalCheckpoint logic

* Cleanup - file organiation

* small nits

* Update tickUpper if new tick is initialized before it

* change outbid checkpoint to outbid block

* edit interface

* Fix conflicts after merge

* Add tests and simpify logic

* use inline assembly for zero case

* fix tickUpper calculations

* Clean up logic and natspec

* Change to getLowerTickForPrice

* Fix clearing logic for price under tickLower

* Fix clearing price calculation

* fix iteration

* fix clearing logic

* Fix tick initialization

* feedback: don't pass in clearing price to update checkpoint

* Change to lastInitializedTickId

* Use prices as keys for ticks

* Use id as price / tick spacing, add back in prev pointer

* Remove headTickId and add max next

* forge fmt

* remove tickLib

* rm

* Remove prev pointer

* rename

* rename

* rename withdraw* to exit*

* finish rename

* comment

* Move to Q96, remove tickSpacing usage

* remove tickId

* Fix accounting logic  (#25)

* save state

* save

* save state

* exact out bug

* Remove logs

* remove comment

* rename

* Remove require in auction

* remove unused var

* refactor

* merge conflict

* fix stack too deep

* merge conflicts

* Before moving to mps * price

* Move by Q96 * 2

* save

* Fix tests besides 1

* feedback: refactor tickStorage to use derived ids from price (#21)

* Refactor to upper tick

* Fix conflicts after merge

* Add tests and simpify logic

* use inline assembly for zero case

* fix tickUpper calculations

* Clean up logic and natspec

* Change to getLowerTickForPrice

* fix iteration

* Change to lastInitializedTickId

* Use prices as keys for ticks

* Use id as price / tick spacing, add back in prev pointer

* Remove headTickId and add max next

* forge fmt

* remove tickLib

* rm

* Remove prev pointer

* rename

* rename

* comment

* merge conflict

* fix stack too deep

* gas

* Still have off by 1 but clean up

* gas

* Comments

* forge fmt

* Remove unused errors

* Fix test

* Remove unused import

* fix tests

+++ src/interfaces/IValidationHook.sol
@@ -13,7 +13,7 @@ interface IValidationHook {
     /// @param sender The sender of the bid
     /// @param hookData Additional data to pass to the hook required for validation
     function validate(
-        uint128 maxPrice,
+        uint256 maxPrice,
         bool exactIn,
         uint256 amount,
         address owner,
